% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.component_post}
\alias{.component_post}
\title{Posterior summaries (mean + CI) of one component on train or test design.}
\usage{
.component_post(fit, k, test = TRUE, probs = c(0.025, 0.975))
}
\description{
Posterior summaries (mean + CI) of one component on train or test design.
}
\keyword{internal}
