% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartman.R
\name{mvbcf_trees}
\alias{mvbcf_trees}
\title{Build a bartMan trees_data object from a component's forest}
\usage{
mvbcf_trees(fit, component = "mu")
}
\arguments{
\item{fit}{A fit object.}

\item{component}{Component label/index (default "mu").}
}
\value{
A bartMan trees data object.
}
\description{
Lets you use native \pkg{bartMan} plotting (e.g. \code{plotTrees},
\code{splitDensity}) on any component forest of the multi-treatment MVBCF.
}
