% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvbcfMT-package.R
\docType{package}
\name{mvbcfMT-package}
\alias{mvbcfMT}
\alias{mvbcfMT-package}
\title{mvbcfMT: Multi-Treatment Multivariate Bayesian Causal Forests}
\description{
Extends the single-treatment multivariate BCF (mvbcfMET) to estimate the joint
causal effect of \strong{two or more} crossed binary treatments on several
outcomes at once. The response surface is decomposed over the treatment
lattice into a prognostic forest \eqn{\mu} plus one forest per (main or
interaction) effect \eqn{\tau_S}, all sharing a residual covariance \eqn{\Sigma}.
A single unified C++ engine (\code{fast_bart_multi}) samples every component.
}
\author{
\strong{Maintainer}: Danilo A. Sarti \email{danilo.stats@gmail.com} (Alternative e-mail: danilosarti@rcsi.ie) [copyright holder]

}
\keyword{internal}
