% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsup.R
\name{plot_effect_forest}
\alias{plot_effect_forest}
\title{Forest plot of the average component effects (with 95\% CrI)}
\usage{
plot_effect_forest(fit, response = fit$responses[1], test = FALSE)
}
\arguments{
\item{fit}{A fit object.}

\item{response}{Which response.}

\item{test}{Use test rows/grid (default FALSE = training rows).}
}
\value{
A ggplot.
}
\description{
Context-agnostic summary: the posterior mean and 95\% credible interval of the
average effect of every component, for one response. Works in any domain.
}
